up file

+++ src/main/java/nhuconghaui/webdidong/formatter/RoleFormatter.java
@@ -1,5 +1,7 @@
 package nhuconghaui.webdidong.formatter;
 
+import nhuconghaui.webdidong.model.Role;
+import nhuconghaui.webdidong.service.RoleService;
 import org.springframework.format.Formatter;
 
 import java.text.ParseException;
@@ -14,7 +16,7 @@ public class RoleFormatter implements Formatter<Role> {
 
     @Override
     public Role parse(String text, Locale locale) throws ParseException {
-        return roleService.findById(Integer.valueOf(text));
+        return roleService.findById(Long.valueOf(text));
     }
 
     @Override
